deploy/pyproject.toml
[project]
name = "mips-emulator"
version = "0.1.0"
dependencies = [ "flask==3.1.0", "flask-cors==5.0.1", "psycopg2-binary==2.9.10", "werkzeug==3.1.3", "python-dotenv==1.0.1", "openpyxl==3.1.5", "requests==2.32.3" ]
requires-python = ">=3.12"